% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{print hello world by default}
\usage{
init(greeting = TRUE)
}
\arguments{
\item{boolean}{greeting}
}
\value{
void
}
\description{
print hello world by default
}
\examples{
init()
init(FALSE)
}
\seealso{
\code{\link{nchar}} which this function wraps
}
